Move array mapper classes to core and make them internal. Free SQL arrays on statement cleanup.
