+++ core/src/main/java/org/jdbi/v3/core/Jdbi.java
@@ -32,7 +32,6 @@ import javax.sql.DataSource;
 import org.jdbi.v3.core.argument.ArgumentFactory;
 import org.jdbi.v3.core.argument.ArrayElementMapper;
 import org.jdbi.v3.core.argument.ArrayElementMapperFactory;
-import org.jdbi.v3.core.argument.InferredArrayElementMapperFactory;
 import org.jdbi.v3.core.collector.CollectorFactory;
 import org.jdbi.v3.core.exception.UnableToObtainConnectionException;
 import org.jdbi.v3.core.extension.ExtensionCallback;
