+++ core/src/main/java/org/jdbi/v3/core/SqlArrayMapperFactory.java
@@ -11,28 +11,27 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.v3.core.argument;
+package org.jdbi.v3.core;
 
 import java.lang.reflect.Type;
 import java.util.List;
 import java.util.Optional;
 
-import org.jdbi.v3.core.StatementContext;
+import org.jdbi.v3.core.mapper.ColumnMapper;
+import org.jdbi.v3.core.mapper.ColumnMapperFactory;
 import org.jdbi.v3.core.util.GenericTypes;
 
-public class SqlArrayArgumentFactory implements ArgumentFactory {
+class SqlArrayMapperFactory implements ColumnMapperFactory {
     @Override
-    public Optional<Argument> build(Type type, Object value, StatementContext ctx) {
-        Class<?> erasedType = GenericTypes.getErasedType(type);
-        if (erasedType.isArray()) {
-            Class<?> elementType = erasedType.getComponentType();
-            return ctx.findArrayElementMapperFor(elementType)
-                    .map(mapper -> new SqlArrayArgument(mapper, value));
+    public Optional<ColumnMapper<?>> build(Type type, StatementContext ctx) {
+        final Class<?> clazz = GenericTypes.getErasedType(type);
+        if (clazz.isArray()) {
+            return Optional.of(new ArrayColumnMapper(clazz.getComponentType()));
         }
-        if (List.class.isAssignableFrom(erasedType)) {
+        if (List.class.isAssignableFrom(clazz)) {
             return GenericTypes.findGenericParameter(type, List.class)
-                    .flatMap(ctx::findArrayElementMapperFor)
-                    .map(mapper -> new SqlArrayArgument(mapper, (List) value));
+                    .flatMap(elementType -> ctx.findColumnMapperFor(elementType))
+                    .map(ListColumnMapper::new);
         }
         return Optional.empty();
     }
